+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Changed
+ - Fixed some bad assumptions when generating console abbreviations
+
 ## [4.2.0] - 2020-03-20
 ### Added
  - ICD console supports setting/clearing root with '/' command